+++ src/main/java/chen/binflare/dao/UserDAO.java
@@ -2,5 +2,7 @@ package chen.binflare.dao;
 
 import chen.binflare.entity.UserDO;
 
-public interface UserDAO extends BaseRepository<UserDO, Long> {
+import java.math.BigInteger;
+
+public interface UserDAO extends BaseRepository<UserDO, BigInteger> {
 }
